+++ spec/features/draws/draw_suite_update_spec.rb
@@ -6,7 +6,7 @@ RSpec.feature 'Draw suite update' do
   let(:draw) { create(:draw_with_members, suites_count: 1) }
   let!(:removed_suite) { draw.suites.first }
   let(:other_draw) { create(:draw_with_members, suites_count: 1) }
-  let!(:other_draw_suite) { other_draw.suites.first }
+  let!(:other_draw_suite) { SuiteDecorator.new(other_draw.suites.first) }
   let!(:undrawn_suite) { create(:suite_with_rooms, rooms_count: 1) }
 
   before { log_in create(:admin) }